pyproject.toml
[tool.poetry]
name = "reddit"
version = "0.1.0"
description = ""
authors = ["pymasters.pl"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.4"
python-decouple = "^3.8"
pytest-django = "^4.8.0"
pillow = "^10.3.0"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2024.2"
django-timesince = "^2021.5.27"
six = "^1.16.0"
pytest = "^8.2.2"
faker = "^25.9.1"
pycountry = "^24.6.1"
djangorestframework = "^3.15.2"
freezegun = "^1.5.1"
pytest-xdist = "^3.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
ruff = "^0.4.5"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["scripts/", "manage.py", "asgi.py", "wsgi.py", "*/migrations/"]
line-length = 120

indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D104", "D100", "D103", "D211", "D213", "D101", "D102", "D105", "D106", "D107", "D203", "ISC001", "COM812"]

[tool.ruff.format]
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"**/tests/**/*.py" = [
    # at least this three should be fine in tests:
    "S101", # asserts allowed in tests...
    "ARG", # Unused function args -> fixtures nevertheless are functionally relevant...
    "FBT", # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
    "RUF001", # Ruff had a problem with "’" in text: The two password fields didn’t match.
    "PLR2004", # Ruff is pushing for constants instead of ints.
]
